Move utils/create-stig-overlay.py to ssg/_utils_overlay.py
